+++ src/network/visualizeruploader.h
@@ -24,7 +24,7 @@ struct ShotMetadata {
     QString grinderBurrs;
     QString grinderSetting;
     qint64 equipmentId = 0; // Equipment package (add-equipment-packages); 0 = none
-    int rpm = 0;            // Grinder rpm dial-in; 0 = unset
+    qint64 rpm = 0;         // Grinder rpm dial-in; 0 = unset
     double beanWeight = 0;  // Dose weight in grams
     double drinkWeight = 0; // Output weight in grams
     double drinkTds = 0;
